Rename src/misc to src/command_wrapper for clarity
