--- src/misc.rs
@@ -1,1 +1,0 @@
-pub mod command_wrapper;
